권한 검사 로직을 utils/auth.ts로 분리 및 리팩토링
- getRequiredAuth 함수를 새로운 파일로 이동하여 코드 구조 개선
- middleware.ts에서 불필요한 코드 제거로 가독성 향상
